+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.25.1] - 2023-04-26
+
+### Changed
+
+- Throttle registration
+- Reorganize routes
+
 ## [1.25.0] - 2023-04-25
 
 ### Changed